requirements.txt
fastapi
uvicorn[standard]
python-dotenv
requests
pinecone
sentence-transformers
torch
openai
PyMuPDF
langchain
tqdm